--- .dockerignore
@@ -1,4 +1,0 @@
-*.yaml
-*.md
-*.lock
-*.json
