+++ .github/workflows/docker-image.yml
@@ -38,7 +38,8 @@ jobs:
         name: Build and push
         uses: docker/build-push-action@v3
         with:
-          context: .
+          file: docker/Dockerfile
+          context: src/
           push: true
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
