+++ src/classes/bot.py
@@ -4,15 +4,14 @@ import logging
 import os
 
 import discord
-from discord.ext import commands
-from motor import motor_asyncio
-
 from classes.config import ConfigList
 from commons.helpers import list_module
 from commons.mongoIO import mongoIO
 from commons.osu.osuapiwrap import osuAPI
 from commons.osu.osuhelpers import osuHelper
 from commons.osu.ppwrap import ppAPI
+from discord.ext import commands
+from motor import motor_asyncio
 
 logger = logging.getLogger()
