+++ src/classes/embeds/OsuListEmbed.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from typing import List
 
-import discord
-
 import classes.osu as osu
+import discord
 
 
 class OsuListEmbed(discord.Embed):
