+++ src/classes/embeds/OsuProfileEmbed.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 from datetime import datetime
 
-import discord
-
 import classes.osu as osu
+import discord
 
 
 class OsuProfileEmbed(discord.Embed):
