+++ src/classes/enums/animals.py
