+++ src/classes/exceptions.py
