+++ src/classes/osu/Beatmap.py
