+++ src/classes/osu/Mods.py
