+++ src/classes/osu/Score.py
