+++ src/classes/osu/User.py
