+++ src/classes/osu/enums/Rank.py
