+++ src/classes/osu/enums/__init__.py
