+++ src/cogs/fun.py
