+++ src/cogs/settings.py
@@ -4,9 +4,8 @@ import time
 from datetime import datetime
 
 import discord
-from discord.ext import commands
-
 from commons import checks
+from discord.ext import commands
 
 disableKeywords = ["none", "disable", "off", "disabled"]
