+++ src/commons/osu/osuapiwrap.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from typing import List
 
 import aiohttp
-
 import classes.osu as osu
 from classes.exceptions import OsuAPIError
