+++ src/commons/osu/ppwrap.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import aiohttp
-
 import classes.osu as osu
