+++ src/run.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import logging
 
 import sentry_sdk
-
 from classes.bot import Sunny
 
 logger = logging.getLogger()
